+++ rustfst/src/fst_traits/mutable_fst.rs
@@ -207,13 +207,6 @@ pub trait MutableFst: Fst + for<'a> MutableArcIterator<'a> {
     fn pop_arcs(&mut self, source: StateId) -> Fallible<Vec<Arc<Self::W>>>;
     unsafe fn pop_arcs_unchecked(&mut self, source: StateId) -> Vec<Arc<Self::W>>;
 
-    /// Reserve space for storing enough arcs leaving a state.
-    fn reserve_arcs(&mut self, source: StateId, additional: usize) -> Fallible<()>;
-    unsafe fn reserve_arcs_unchecked(&mut self, source: StateId, additional: usize);
-
-    /// Reserve space for storing enough states.
-    fn reserve_states(&mut self, additional: usize);
-
     /// Retrieves a mutable reference to the final weight of a state (if the state is a final one).
     fn final_weight_mut(
         &mut self,